Refactor config command integration tests
*  Refactor config command integration tests
   to use common base classes.
