+++ tests/integration/actions/config/test_welcome_interactive_specified_config.py
@@ -2,12 +2,13 @@
 """
 import pytest
 
-from .base import CONFIG_FIXTURE, add_indicies
-from .base import step_id
+from .base import CONFIG_FIXTURE
 from .base import BaseClass
-from .base import Command
-from .base import Step
 
+from ..._interactions import add_indicies
+from ..._interactions import Command
+from ..._interactions import Step
+from ..._interactions import step_id
 
 CLI = Command(execution_environment=False).join()
